requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.0
colorama
art
simple_term_menu
